// main/api/graphQlApi/models.ts
export type Username = string;

/** A name must neither contain whitespace nor exceed 30 characters. */
export type Name = string;

/** A user's bio which cannot exceed 2,500 characters, and uses CommonMark. */
export type Bio = string;
// ...
export interface Account extends AccountData {
    readonly id: number;
    readonly username: Username;
    readonly emailAddress: string;
    readonly firstName: Name;
    readonly lastName: Name;
    readonly bio: Bio;
}

export interface AccountsConnection {
    readonly edges: AccountEdge[];
    readonly pageInfo: PageInfo;
}

export interface AccountEdge {
    readonly node: Account;
    readonly cursor: Cursor;
}

export interface PageInfo {
    readonly hasNextPage: Boolean;
    readonly hasPreviousPage: Boolean;
    readonly startCursor?: Cursor;
    readonly endCursor?: Cursor;
}

export interface AccountData {
    readonly id: number;
    readonly username: Username;
    readonly emailAddress: string;
    readonly firstName: Name;
    readonly lastName: Name;
    readonly bio: Bio;
}